WebSockets/tests/submissions/Opera/opening-handshake/002.html

<!doctype html>
<title>WebSockets: valid handshake</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../constants.js></script>
<div id=log></div>
<script>
setTestTimeout(10000, 'got no open event');
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':8007/echo');
  ws.onopen = function(e) {
    debug(e);
    ws.onclose = function(e) {
      debug(e);
      assert_equals(e.wasClean, true);
      ws.onclose = assertUnreached;
      setTestTimeout(null);
      setTimeout(end, 50);
    }
    ws.close();
  }
  ws.onerror = ws.onmessage = ws.onclose = assertUnreached;
});
</script>
